requirements.txt
beautifulsoup4
numpy>1.23.2
pandas>1.5.2
protobuf>3.20.3
Requests>2.32.3
ScraperFC
selenium
tqdm
webdriver_manager
google-cloud-bigquery
pandas-gbq
psutil
rapidfuzz